Use size_t for length and count variables to fix C4267

+++ src/romaji.c
@@ -437,7 +437,7 @@ romaji_convert2(romaji *object, const unsigned char *string,
             {
                 if (string[i])
                 {
-                    stop = WORDBUF_LEN(buf);
+                    stop = (int)WORDBUF_LEN(buf);
                     wordbuf_cat(buf, &string[i]);
                 }
                 break;
